+++ test/entry.test.js
@@ -9,12 +9,40 @@ import gpt from './responses/openai/text.json';
 import claude from './responses/anthropic/text.json';
 
 describe('entry', () => {
+  it('should error on bad setup', async () => {
+    expect(() => {
+      new Client();
+    }).toThrow('No platform specified.');
+
+    expect(() => {
+      new Client({
+        platform: 'openai',
+      });
+    }).toThrow('No templates directory specified.');
+
+    expect(() => {
+      new Client({
+        platform: 'openai',
+        templates: 'test',
+      });
+    }).toThrow('No API key specified.');
+
+    expect(() => {
+      new Client({
+        platform: 'foobar',
+        templates: 'test',
+        apiKey: 'test-key',
+      });
+    }).toThrow('Unknown platform "foobar".');
+  });
+
   it('should succeed for openai', async () => {
     setGptResponse(gpt);
 
     const client = new Client({
       platform: 'gpt',
       templates: path.join(__dirname, './templates'),
+      apiKey: 'test-key',
     });
     const result = await client.prompt({
       file: 'stocks',
@@ -36,7 +64,9 @@ describe('entry', () => {
     const client = new Client({
       platform: 'claude',
       templates: path.join(__dirname, './templates'),
+      apiKey: 'test-key',
     });
+
     const result = await client.prompt({
       file: 'stocks',
       output: 'json',